pyproject.toml
[tool.poetry]
name = "icfpc-2024"
version = "1.0"
authors = ["DNIWE :: a"]
description = "Solver for ICFPC 2024"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
